renamed the rest of the decorator modules
