+++ src/earthkit/utils/decorators/_format_handlers.py
